Introduce indexables-head-action for consistency between routes and cleaner classes
